+++ changelog.md
@@ -1,7 +1,7 @@
 # Summary of user-visible changes
 
 ## ??? / ???
-
+* Add bit-wise operators `>>`, `<<`, `|`, `&`, `bnot`, and `xor`
 * Add support for __fennelview metamethod for custom serialization
 * Fix a bug where `dofile` would report the wrong filename
 * Fix bug causing failing `include` of Lua modules that lack a trailing newline (#234)