+++ changelog.md
@@ -5,10 +5,10 @@
 * Add bit-wise operators `>>`, `<<`, `|`, `&`, `bnot`, and `xor`
 * Friendlier compiler/parse error messages with suggestions
 * Omit compiler internal stack traces by default unless `FENNEL_DEBUG=trace`
-* Add support for __fennelview metamethod for custom serialization
+* Add support for `__fennelview` metamethod for custom serialization
 * Fix a bug where `dofile` would report the wrong filename
 * Fix bug causing failing `include` of Lua modules that lack a trailing newline (#234)
-- Introduce `limit-values` and `limit-args` macros (#246)
+* Introduce `pick-values` and `pick-args` macros (as `limit-*`: #246, as `pick-*`: #256)
 
 ## 0.3.2 / 2020-01-14